+++ src/classes/main-worker.ts
@@ -1,0 +1,8 @@
+/**
+ * Worker Thread wrapper for sandboxing
+ *
+ */
+import { parentPort } from 'worker_threads';
+import mainBase from './main-base';
+
+mainBase(async (msg: any) => parentPort.postMessage(msg), parentPort);